IMPROVEMENT added extra logging about unspent outputs when something is not right

+++ app/database/seeds/BitcoinAddressesSeeder.php
@@ -4,7 +4,11 @@ class BitcoinAddressesSeeder extends Seeder {
 
 	public function run()
 	{
-		DB::table('addresses')->delete();
+		// not to delete in production just in case lol
+		if ( App::environment('production') )
+		{
+			DB::table('addresses')->delete();
+		}
 
 		Address::create(['address' => 'mtqSqRAGB7EuPtgCwvMtX74S2tvTapDWD6', 'user_id' => 1]);
 		Address::create(['address' => 'n21cjTZa59QcMBXFvoKx2WoRotBV9mErnJ', 'user_id' => 1]);
