+++ app/database/seeds/CryptoTypeTableSeeder.php
@@ -4,7 +4,11 @@ class CryptoTypeTableSeeder extends Seeder {
 
 	public function run()
 	{
-		DB::table('crypto_types')->delete();
+		// not to delete in production just in case lol
+		if ( App::environment('production') )
+		{
+			DB::table('crypto_types')->delete();
+		}
 
 		CryptoType::create(['crypto_type' => 'BTC']);
 	}
